+++ tests/integration/test_metadata.rs
@@ -1,6 +1,6 @@
 // Copyright © 2024 Pathway
 
-use super::helpers::read_data_from_reader;
+use super::helpers::{new_csv_filesystem_reader, new_filesystem_reader, read_data_from_reader};
 
 use std::collections::HashMap;
 
@@ -8,9 +8,7 @@ use pathway_engine::connectors::data_format::{
     DsvParser, DsvSettings, IdentityParser, InnerSchemaField, JsonLinesParser, KeyGenerationPolicy,
     ParsedEvent,
 };
-use pathway_engine::connectors::data_storage::{
-    new_csv_filesystem_reader, new_filesystem_reader, ConnectorMode, ReadMethod,
-};
+use pathway_engine::connectors::data_storage::{ConnectorMode, ReadMethod};
 use pathway_engine::connectors::SessionType;
 use pathway_engine::engine::{Type, Value};